TPJoas/requirements.txt
matplotlib
numpy
pillow
SciPy 